Finalize impl
