--- distage/distage-extension-plugins/src/main/scala/izumi/distage/plugins/StaticPluginScanner.scala
@@ -1,7 +1,0 @@
-package izumi.distage.plugins
-
-import scala.language.experimental.macros
-
-object StaticPluginScanner {
-  def staticallyAvailablePlugins(pluginsPackage: String): List[PluginBase] = macro StaticPluginScannerMacro.staticallyAvailablePlugins
-}
