+++ distage/distage-extension-plugins/src/main/scala/izumi/distage/plugins/load/PluginLoaderDefaultImpl.scala
@@ -62,7 +62,7 @@ object PluginLoaderDefaultImpl {
       .map(_.acceptPackages(enabledPackages: _*))
       .map(_.acceptClasses(whitelistClasses :+ base: _*))
       .map(_.rejectPackages(disabledPackages: _*))
-      .map(_.enableClassInfo())
+      .map(_.enableExternalClasses())
       .map(if (debug) _.verbose() else identity)
       .map(_.scan())
       .get
