+++ distage/distage-extension-plugins/src/test/scala/com/github/pshirshov/test/PluginLoaderTest.scala
@@ -1,6 +1,6 @@
 package com.github.pshirshov.test
 
-import com.github.pshirshov.test.plugins.{DependingPlugin, EmptyTestPlugin, StaticTestPlugin}
+import com.github.pshirshov.test.plugins.{DependingPlugin, EmptyTestPlugin, ObjectTestPlugin, StaticTestPlugin}
 import distage.plugins.PluginLoader
 import izumi.distage.plugins.PluginConfig
 import org.scalatest.wordspec.AnyWordSpec
@@ -8,7 +8,7 @@ import org.scalatest.wordspec.AnyWordSpec
 class PluginLoaderTest extends AnyWordSpec {
   "Load plugins list at runtime time" in {
     val plugins = PluginLoader().load(PluginConfig.packages(Seq("com.github.pshirshov.test.plugins")))
-    assert(plugins.size == 5)
+    assert(plugins.size == 6)
     assert(
       plugins.map(_.getClass).toSet == Set(
         EmptyTestPlugin.getClass,
@@ -16,6 +16,7 @@ class PluginLoaderTest extends AnyWordSpec {
         classOf[DependingPlugin],
         classOf[DependingPlugin.NestedDoublePlugin],
         DependingPlugin.NestedDoublePlugin.getClass,
+        ObjectTestPlugin.getClass,
       )
     )
   }
