+++ distage/distage-extension-plugins/src/test/scala/com/github/pshirshov/test/plugins/EmptyTestPlugin.scala
@@ -1,0 +1,5 @@
+package com.github.pshirshov.test.plugins
+
+import com.github.pshirshov.test.sneaky.SneakyPlugin
+
+object EmptyTestPlugin extends SneakyPlugin
