+++ distage/distage-extension-plugins/src/test/scala/com/github/pshirshov/test/plugins/ObjectTestPlugin.scala
@@ -1,0 +1,7 @@
+package com.github.pshirshov.test.plugins
+
+import com.github.pshirshov.test.sneaky.SneakyPlugin
+
+object ObjectTestPlugin extends SneakyPlugin {
+  make[TestDep3]
+}
