+++ distage/distage-extension-plugins/src/test/scala/com/github/pshirshov/test/plugins/StaticTestPlugin.scala
@@ -1,17 +1,14 @@
 package com.github.pshirshov.test.plugins
 
 import com.github.pshirshov.test.sneaky.SneakyPlugin
-import distage.{Axis, ModuleDef}
+import distage.Axis
 import izumi.distage.config.ConfigModuleDef
 import izumi.fundamentals.platform.build.ExposedTestScope
-import izumi.logstage.api.IzLogger
 
 object Test extends Axis {
   case object X extends AxisValueDef
   case object Y extends AxisValueDef
 }
-
-object EmptyTestPlugin extends SneakyPlugin
 
 @ExposedTestScope
 class StaticTestPlugin extends SneakyPlugin with ConfigModuleDef {
@@ -20,40 +17,3 @@ class StaticTestPlugin extends SneakyPlugin with ConfigModuleDef {
   make[TestService]
   make[TestConf].fromConfig("test.testconf")
 }
-
-@ExposedTestScope
-class DependingPlugin extends SneakyPlugin {
-  include(DependingPlugin.module)
-  make[TestDepending]
-}
-
-@ExposedTestScope
-object DependingPlugin {
-  val module: ModuleDef = new ModuleDef {
-    // FIXME: make[] redundant generator ???
-    make[Unit].from((_: TestDep, _: TestService) => ()) // FIXME: Provider equals generates conflicting exception on merge
-  }
-  class NestedDoublePlugin extends SneakyPlugin
-  object NestedDoublePlugin extends SneakyPlugin
-}
-
-@ExposedTestScope
-class TestDepending(val testDep: TestDep, val testService: TestService)
-
-@ExposedTestScope
-trait TestDep
-
-@ExposedTestScope
-class TestDep1 extends TestDep
-
-@ExposedTestScope
-class TestDep2 extends TestDep
-
-@ExposedTestScope
-class TestService(
-  val testConf: TestConf,
-  val log: IzLogger,
-)
-
-@ExposedTestScope
-final case class TestConf(int: Int)
