+++ distage/distage-extension-plugins/src/test/scala/com/github/pshirshov/test/sneaky/SneakyPlugin.scala
@@ -1,11 +1,11 @@
 package com.github.pshirshov.test.sneaky
 
 import izumi.distage.model.definition.ModuleDef
-import izumi.distage.plugins.PluginBase
+import izumi.distage.plugins.{ForcedRecompilationToken, PluginBase}
 import izumi.fundamentals.platform.build.ExposedTestScope
 
 /**
   * This is just to verify that plugin enumerator picks up transitively inherited plugins
   */
 @ExposedTestScope
-trait SneakyPlugin extends PluginBase with ModuleDef
+abstract class SneakyPlugin[T <: String](implicit val ev: ForcedRecompilationToken[T]) extends PluginBase with ModuleDef
